Rename GRPCClient to GRPCClientConnection
