+++ Sources/SwiftGRPCNIO/ClientCalls/BidirectionalStreamingClientCall.swift
@@ -29,11 +29,11 @@ import NIO
 public class BidirectionalStreamingClientCall<RequestMessage: Message, ResponseMessage: Message>: BaseClientCall<RequestMessage, ResponseMessage>, StreamingRequestClientCall {
   private var messageQueue: EventLoopFuture<Void>
 
-  public init(client: GRPCClient, path: String, callOptions: CallOptions, handler: @escaping (ResponseMessage) -> Void) {
-    self.messageQueue = client.channel.eventLoop.makeSucceededFuture(())
-    super.init(client: client, path: path, callOptions: callOptions, responseObserver: .callback(handler))
+  public init(connection: GRPCClientConnection, path: String, callOptions: CallOptions, handler: @escaping (ResponseMessage) -> Void) {
+    self.messageQueue = connection.channel.eventLoop.makeSucceededFuture(())
+    super.init(connection: connection, path: path, callOptions: callOptions, responseObserver: .callback(handler))
 
-    let requestHead = self.makeRequestHead(path: path, host: client.host, callOptions: callOptions)
+    let requestHead = self.makeRequestHead(path: path, host: connection.host, callOptions: callOptions)
     self.messageQueue = self.messageQueue.flatMap { self.sendHead(requestHead) }
   }
