+++ Sources/protoc-gen-swiftgrpc/Generator-Client.swift
@@ -446,21 +446,21 @@ extension Generator {
   }
 
   private func printNIOServiceClientImplementation() {
-    println("\(access) final class \(serviceClassName)Client: GRPCServiceClient, \(serviceClassName) {")
+    println("\(access) final class \(serviceClassName)Client: GRPCClient, \(serviceClassName) {")
     indent()
-    println("\(access) let client: GRPCClient")
+    println("\(access) let connection: GRPCClientConnection")
     println("\(access) let service = \"\(servicePath)\"")
     println("\(access) var defaultCallOptions: CallOptions")
     println()
     println("/// Creates a client for the \(servicePath) service.")
     println("///")
     printParameters()
-    println("///   - client: `GRPCClient` with a connection to the service host.")
+    println("///   - connection: `GRPCClientConnection` to the service host.")
     println("///   - defaultCallOptions: Options to use for each service call if the user doesn't provide them. Defaults to `client.defaultCallOptions`.")
-    println("\(access) init(client: GRPCClient, defaultCallOptions: CallOptions? = nil) {")
+    println("\(access) init(connection: GRPCClientConnection, defaultCallOptions: CallOptions? = nil) {")
     indent()
-    println("self.client = client")
-    println("self.defaultCallOptions = defaultCallOptions ?? client.defaultCallOptions")
+    println("self.connection = connection")
+    println("self.defaultCallOptions = defaultCallOptions ?? connection.defaultCallOptions")
     outdent()
     println("}")
     println()
@@ -477,7 +477,7 @@ extension Generator {
         println("/// - Returns: A `UnaryClientCall` with futures for the metadata, status and response.")
         println("\(access) func \(methodFunctionName)(_ request: \(methodInputName), callOptions: CallOptions? = nil) -> UnaryClientCall<\(methodInputName), \(methodOutputName)> {")
         indent()
-        println("return UnaryClientCall(client: client, path: path(forMethod: \"\(method.name)\"), request: request, callOptions: callOptions ?? self.defaultCallOptions)")
+        println("return UnaryClientCall(connection: self.connection, path: self.path(forMethod: \"\(method.name)\"), request: request, callOptions: callOptions ?? self.defaultCallOptions)")
         outdent()
         println("}")
 
@@ -491,7 +491,7 @@ extension Generator {
         println("/// - Returns: A `ServerStreamingClientCall` with futures for the metadata and status.")
         println("\(access) func \(methodFunctionName)(_ request: \(methodInputName), callOptions: CallOptions? = nil, handler: @escaping (\(methodOutputName)) -> Void) -> ServerStreamingClientCall<\(methodInputName), \(methodOutputName)> {")
         indent()
-        println("return ServerStreamingClientCall(client: client, path: path(forMethod: \"\(method.name)\"), request: request, callOptions: callOptions ?? self.defaultCallOptions, handler: handler)")
+        println("return ServerStreamingClientCall(connection: self.connection, path: self.path(forMethod: \"\(method.name)\"), request: request, callOptions: callOptions ?? self.defaultCallOptions, handler: handler)")
         outdent()
         println("}")
 
@@ -505,7 +505,7 @@ extension Generator {
         println("/// - Returns: A `ClientStreamingClientCall` with futures for the metadata, status and response.")
         println("\(access) func \(methodFunctionName)(callOptions: CallOptions? = nil) -> ClientStreamingClientCall<\(methodInputName), \(methodOutputName)> {")
         indent()
-        println("return ClientStreamingClientCall(client: client, path: path(forMethod: \"\(method.name)\"), callOptions: callOptions ?? self.defaultCallOptions)")
+        println("return ClientStreamingClientCall(connection: self.connection, path: self.path(forMethod: \"\(method.name)\"), callOptions: callOptions ?? self.defaultCallOptions)")
         outdent()
         println("}")
 
@@ -520,7 +520,7 @@ extension Generator {
         println("/// - Returns: A `ClientStreamingClientCall` with futures for the metadata and status.")
         println("\(access) func \(methodFunctionName)(callOptions: CallOptions? = nil, handler: @escaping (\(methodOutputName)) -> Void) -> BidirectionalStreamingClientCall<\(methodInputName), \(methodOutputName)> {")
         indent()
-        println("return BidirectionalStreamingClientCall(client: client, path: path(forMethod: \"\(method.name)\"), callOptions: callOptions ?? self.defaultCallOptions, handler: handler)")
+        println("return BidirectionalStreamingClientCall(connection: self.connection, path: self.path(forMethod: \"\(method.name)\"), callOptions: callOptions ?? self.defaultCallOptions, handler: handler)")
         outdent()
         println("}")
       }
